test: 예외 사항 테스트 추가 및 기존 테스트에서 예외 사항 테스트 분리
